fix: bidirectional drift guards, shared telemetry test fixtures, validation API trims (#38013)
